Consolidate shared modules
