+++ functions/common/datastore.js
