+++ functions/common/kml-writer.js
